+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 1.0.0a19 (unreleased)
 ---------------------
 
-- Nothing changed yet.
+- Fix conflict error retries and make tests for it
+  [vangheem]
 
 
 1.0.0a18 (2017-05-08)